Improved Multi-Controller Reliability
Replaced the call to libusb_open_device_with_vid_pid() with libusb_open(). The former should not be used outside of a test application and has limitations such as only being able to open the first device with a matching vid/pid. This caused issues with multiple identical controllers, where only the first would be opened and the second would fail.