Added Deadzone Functionality
Deadzone and adjustment button combination LT + RT + (LS or RS) + DPAD
UP/DN
has been added per controller. Current deadzone is displayed in the
context menu. The deadzone is implemented by mapping the remaining range
of the axis back to the range of 0 to SHRT_MAX. Left and Right analog
sticks can be adjusted individually and is determined by which analog
stick
is pressed down during the deadzone adjustment button combination.

TODO: It would be nice if the deadzone would be remembered by hardware
serial number and stored in the registry or elsewhere.