Added left / right trigger deadzone & adjustment
Similarly to the analog stick deadzone adjustment, the trigger deadzone
is adjusted by a unique button combination (LT/RT)+LS+RT+D-Pad Up/Dn.
The current values are displayed in the context menu. I implemented this
using the same 2-D (XY) deadZoneCalc() function for the analog sticks
with one of the dimensions ignored by passing it a NULL *y_outand 0 for
the input y value.